Fixed sheep move from wolf
Dogs run away from you if you are wolf. But if they see you as a wolf you can't run away if you transform to sheep. Amongus sound for killing sheep. Sheep can be attracted by the player with a cooldown. UI shows when the player is detected as a wolf